fix(data-plane-v2): every path query returned zero, and the trace endpoint that would have shown it was down
Round 8 gated the mined lexical class at recall 1.0000 / nDCG 0.9622 and called
it healthy. That number was a poor proxy: all 23 mined queries are snake_case or
SCREAMING_CASE identifiers and **not one contains a `/` or a `.`**, so the gate
only ever exercised the category that worked.

**1. `/` was stripped from every query, and both lexical arms went dark on paths.**

`sanitize_terms` removed `/` as "path syntax". But Postgres' `simple` parser
indexes a file path as ONE lexeme containing its slashes, so stripping the
separators produced a term that cannot match anything indexed. Measured on the
live 1,164-chunk corpus:

| query as sent | Postgres FTS | Quickwit |
| `src/api/mod.rs` (raw) | 10 rows | 153 hits |
| `srcapimod.rs` (what DP2 sent) | 0 | 0 |
| `zzz/qqq/nope.rs` (control) | - | 0 hits |

The control is the point: 153 was real signal, so the stripped form was
discarding genuine hits, not avoiding false ones. Both arms, not just the
Postgres fallback.

Fixed by preserving `/`, verified safe for both consumers rather than assumed:
`websearch_to_tsquery` is total on user input and a path survives OR-joining
intact (`'src/api/mod.rs' | 'retrieval'`), and `build_quickwit_query` wraps every
term in `quote_term` unconditionally so a slash sits inside a quoted phrase. `:`
stays stripped — it is a Quickwit field selector and no measured recall sits
behind keeping it. Two tests deliberately changed direction; both previously
asserted the mangling. Live after deploy: `src/api/mod.rs` went from 0 sparse
candidates to 10.

**2. `GET /v1/retrieval/{trace_id}` returned HTTP 500 for every trace ever written.**

    decoding column "dense_score": Rust type `Option<f32>` (as SQL type FLOAT4)
    is not compatible with SQL type FLOAT8

`TraceCandidateRow` declared four score columns as `f32`; they are `double
precision`, and sqlx type-checks decodes strictly. 5,886 runs and 58,759
candidate rows had been written and none could be read back.

The endpoint is not the real cost. It is the ONLY surface exposing per-arm
attribution — `candidate_count_dense`, `candidate_count_sparse`, per-stage
timings. With it down, arm liveness could only be inferred from the retrieve
response, which carries score fields for dense and sparse alone and nothing for
graph, wiki, visual, keyword, audio or video. That is why arm liveness has been
misread repeatedly in the evidence doc, twice in this round alone: I read
`sparse_score == 0` as "the sparse arm is dark" when it only meant "this
candidate came from an arm with no field to report". Fix the instrument before
trusting a reading from it.

Fixed to `f64`, narrowing to `f32` at the gRPC boundary where the proto declares
`float`.

**Measured and left open, deliberately:** realistic query shapes cost 0.115 nDCG.
The same 23 identifiers wrapped in the words a person would type around them
(`where is X defined and how is it used`) score recall 0.9130 / nDCG 0.8473
against 1.0000 / 0.9622 bare — more than twice the gate's tolerance, with two
identifiers becoming unfindable. Judgments are identical, so the prose is the
whole cause. The AND->OR fix that made multi-word questions work at all has swung
to the opposite failure: `"where" OR "is" OR "internal/jobs/executor.go" OR
"called"` returns 1,517 of ~2,328 chunks with the path in none of the top 5,
while the path alone returns 14. Stopword handling, minimum-should-match and
rare-term boosting are the standard remedies; choosing among them needs
measurement, so this is recorded rather than guessed at.

Verified: 862 workspace tests, fmt and clippy clean, lexical gate re-run
unchanged (+0.0000 on all six metrics).

+++ apps/Data Plane v2/services/retrieval-engine-rs/src/grpc/retrieval_svc.rs
@@ -406,9 +406,16 @@ impl RetrievalService for RetrievalSvc {
                             rank: c.rank,
                             knowledge_id: c.knowledge_id.unwrap_or_default(),
                             document_id: c.document_id.unwrap_or_default(),
-                            dense_score: c.dense_score.unwrap_or(0.0),
-                            sparse_score: c.sparse_score.unwrap_or(0.0),
-                            rerank_score: c.rerank_score.unwrap_or(0.0),
+                            // Narrowed here, at the wire boundary, because the
+                            // proto declares these `float`. The DB columns are
+                            // `double precision` and `TraceCandidateRow` now
+                            // decodes them as `f64` to match — reading them as
+                            // `f32` made every trace fetch fail. Scores are
+                            // ranking signals in [0,1]-ish ranges, so f32 on the
+                            // wire loses nothing that a caller can act on.
+                            dense_score: c.dense_score.unwrap_or(0.0) as f32,
+                            sparse_score: c.sparse_score.unwrap_or(0.0) as f32,
+                            rerank_score: c.rerank_score.unwrap_or(0.0) as f32,
                             source_chunk_ref: String::new(),
                         })
                         .collect(),
